add comments, move name and symbol to base
