+++ tests/token/erc20/ERC20BaseSuite.py
@@ -61,6 +61,20 @@ class ERC20Base:
 
 
     @pytest.mark.asyncio
+    async def test_name(self, contract_factory):
+        erc20, _, _ = contract_factory
+        execution_info = await erc20.name().execute()
+        assert execution_info.result.name == NAME
+
+
+    @pytest.mark.asyncio
+    async def test_symbol(self, contract_factory):
+        erc20, _, _ = contract_factory
+        execution_info = await erc20.symbol().execute()
+        assert execution_info.result.symbol == SYMBOL
+
+
+    @pytest.mark.asyncio
     async def test_decimals(self, contract_factory):
         erc20, _, _ = contract_factory
         execution_info = await erc20.decimals().execute()
